+++ pkg/helpers/helpers.go
@@ -12,6 +12,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/internal/collections"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
 )
 
@@ -142,18 +143,8 @@ func DecodeSnowflakeAccountIdentifier(identifier string) (sdk.AccountIdentifier,
 	}
 }
 
-// TODO(SNOW-1479870): Test
-// MergeMaps takes any number of maps (of the same type) and concatenates them.
-// In case of key collision, the value will be selected from the map that is provided
-// later in the src function parameter.
 func MergeMaps[M ~map[K]V, K comparable, V any](src ...M) M {
-	merged := make(M)
-	for _, m := range src {
-		for k, v := range m {
-			merged[k] = v
-		}
-	}
-	return merged
+	return collections.MergeMaps(src...)
 }
 
 // TODO: use slices.Concat in Go 1.22
