+++ pkg/internal/collections/collection_helpers.go
@@ -22,3 +22,17 @@ func Map[T any, R any](collection []T, mapper func(T) R) []R {
 	}
 	return result
 }
+
+// TODO(SNOW-1479870): Test
+// MergeMaps takes any number of maps (of the same type) and concatenates them.
+// In case of key collision, the value will be selected from the map that is provided
+// later in the src function parameter.
+func MergeMaps[M ~map[K]V, K comparable, V any](src ...M) M {
+	merged := make(M)
+	for _, m := range src {
+		for k, v := range m {
+			merged[k] = v
+		}
+	}
+	return merged
+}
